auth/requirements.txt
email-validator==1.1.2
fastapi==0.89.0
PyJWT==1.7.1
python-decouple==3.3
uvicorn==0.12.3
validators==0.18.1
bcrypt==3.2.0
python-decouple==3.3

